word: simpler. don't prematurely optimize
trying to figure out whether or not mapping has been updated, when dealing
with multiple words in a headline, which need to all share a mapping, started
to get confusing; instead, calculate possibles on the fly every time for now;
noting that there's potential for improvement later, if it's necessary.

My impression is that it might be, if there's an 'attempt to auto solve' fx
on a headline, because every change will require recalculating everything.
but maybe that's inconsequential enough that I don't have to worry about it
now.